Re-organize files for bento build
