+++ extensions/amp-autocomplete/1.0/test/test-amp-autocomplete.js
@@ -1,6 +1,8 @@
 import '../amp-autocomplete';
 import {htmlFor} from '#core/dom/static-template';
+
 import {toggleExperiment} from '#experiments';
+
 import {waitFor} from '#testing/helpers/service';
 
 describes.realWin(
