+++ src/bento/components/bento-autocomplete/1.0/component.js
@@ -1,0 +1,1 @@
+export * from './component/component';
