+++ src/bento/components/bento-autocomplete/1.0/component.type.js
@@ -5,13 +5,9 @@ var BentoAutocompleteDef = {};
 
 /**
  * @typedef {{
- *   exampleProperty: (string|undefined), (DO NOT SUBMIT)
  * }}
  */
 BentoAutocompleteDef.Props;
 
 /** @interface */
-BentoAutocompleteDef.BentoAutocompleteApi = class {
-  /** Example: API method to toggle the component */
-  exampleToggle() {} // DO NOT SUBMIT
-};
+BentoAutocompleteDef.BentoAutocompleteApi = class {};
