+++ src/bento/components/bento-autocomplete/1.0/component/use-autocomplete-binding.ts
@@ -1,6 +1,7 @@
 import {AutocompleteBindingInline} from './autocomplete-binding-inline';
 import {AutocompleteBindingSingle} from './autocomplete-binding-single';
-import {AutocompleteBinding} from './types';
+
+import {AutocompleteBinding} from '../types';
 
 export const useAutocompleteBinding: (trigger?: string) => AutocompleteBinding =
   (trigger) => {
